Fix package powsybl.tutorials to com.powsybl.tutorials
