+++ cgmes/src/main/resources/simplelogger.properties
@@ -1,2 +1,2 @@
-org.slf4j.simpleLogger.log.powsybl.tutorials.cgmes.CgmesMergeTutorial=debug
+org.slf4j.simpleLogger.log.com.powsybl.tutorials.cgmes.CgmesMergeTutorial=debug
 org.slf4j.simpleLogger.log.com.powsybl.cgmes.conversion.Context=error
