+++ count-network-lines/src/main/java/com/powsybl/tutorials/tools/CountNetworkLinesTool.java
@@ -1,0 +1,82 @@
+/**
+ * Copyright (c) 2019, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+package com.powsybl.tutorials.tools;
+
+import java.nio.file.Path;
+
+import org.apache.commons.cli.CommandLine;
+import org.apache.commons.cli.Option;
+import org.apache.commons.cli.Options;
+
+import com.google.auto.service.AutoService;
+import com.powsybl.commons.PowsyblException;
+import com.powsybl.iidm.import_.ImportConfig;
+import com.powsybl.iidm.import_.Importers;
+import com.powsybl.iidm.network.Network;
+import com.powsybl.tools.Command;
+import com.powsybl.tools.Tool;
+import com.powsybl.tools.ToolRunningContext;
+
+/**
+ * @author Abdelsalem Hedhili <abdelsalem.hedhili at rte-france.com>
+ */
+@AutoService(Tool.class)
+public class CountNetworkLinesTool implements Tool {
+
+    private static final String CASE_FILE = "case-file";
+
+    @Override
+    public Command getCommand() {
+        return new Command() {
+
+            @Override
+            public String getName() {
+                return "count-network-lines";
+            }
+
+            @Override
+            public String getTheme() {
+                return "Network";
+            }
+
+            @Override
+            public String getDescription() {
+                return "Count network lines";
+            }
+
+            @Override
+            public Options getOptions() {
+                Options options = new Options();
+                options.addOption(Option.builder().longOpt(CASE_FILE)
+                        .desc("the case path")
+                        .hasArg()
+                        .argName("FILE")
+                        .required()
+                        .build());
+                return options;
+            }
+
+            @Override
+            public String getUsageFooter() {
+                return null;
+            }
+        };
+    }
+
+    @Override
+    public void run(CommandLine line, ToolRunningContext context) {
+        Path caseFile = context.getFileSystem().getPath(line.getOptionValue(CASE_FILE));
+        context.getOutputStream().println("Loading network '" + caseFile + "'");
+        Network network = Importers.loadNetwork(caseFile, context.getShortTimeExecutionComputationManager(), ImportConfig.load(), null);
+        if (network == null) {
+            throw new PowsyblException("Case '" + caseFile + "' not found");
+        }
+        int lineCount = network.getLineCount();
+        context.getOutputStream().println("Network contains '" + lineCount + "' lines");
+    }
+
+}
