+++ csv-importer/pom.xml
@@ -48,7 +48,7 @@
 
     <!-- This property is only needed to run the tutorial easily -->
     <properties>
-      <exec.mainClass>powsybl.tutorials.csv.import_.Main</exec.mainClass>
+      <exec.mainClass>com.powsybl.tutorials.csv.import_.Main</exec.mainClass>
     </properties>
 
 </project>
