+++ downscaling/complete/src/main/resources/simplelogger.properties
@@ -1,1 +1,1 @@
-org.slf4j.simpleLogger.log.powsybl.tutorials.downscaling.Downscaling=debug
+org.slf4j.simpleLogger.log.com.powsybl.tutorials.downscaling.Downscaling=debug
