+++ loadflow/pom.xml
@@ -16,7 +16,7 @@
     <!-- These properties are only needed to run the tutorial from the command line with mvn exec:java -->
     <properties>
         <exec.cleanupDaemonThreads>false</exec.cleanupDaemonThreads>
-        <exec.mainClass>powsybl.tutorials.loadflow.LoadflowTutorial</exec.mainClass>
+        <exec.mainClass>com.powsybl.tutorials.loadflow.LoadflowTutorial</exec.mainClass>
     </properties>
 
     <build>
