+++ merging-view/pom.xml
@@ -15,7 +15,7 @@
     <!-- These properties are only needed to run the tutorial from the command line with mvn exec:java -->
     <properties>
         <exec.cleanupDaemonThreads>false</exec.cleanupDaemonThreads>
-        <exec.mainClass>powsybl.tutorials.mergingview.MergingViewTutorial</exec.mainClass>
+        <exec.mainClass>com.powsybl.tutorials.mergingview.MergingViewTutorial</exec.mainClass>
     </properties>
 
     <build>
